WIP: Refactor
Not sure exactly what is going on here. I guess:

1/ By introducing an "editor" interface between otto and the db,
   the circular dep between DB and Otto can be broken and all the
   commands can be in DB and the main interface can be in DB.

2/ Also, separating out the Editor concept makes Database simpler
   because it doesn't have to track the previous values, open editors,
   etc. itself. This way its lifecycle also becomes easier to understand.

However, once you do this, the sync package breaks because it is
trying to manually run transactions against otto, so it now needs to
also create an editor and do the right things.

I guess one thing I could do is:

Make the relevant APIs in Database public so that I can do the
minimal work to get the tests running again.

Alternately, I could just take the Editor interface and fork the code
and try and build up the high-level interface in new_api_sketch.txt
and see if that works.