Use JSON Patch compatible clear
This is controlled by the version field. The new version is `2` and the
logic should only apply to that version.

When clearing the map we now send
`{"op":"replace","path":"","value":{}}` instead of
`{"op":"remove","path":"/"}` which really should mean remove the key
`""`/

Towards https://github.com/rocicorp/repc/issues/124